requirements.txt
masonite-orm==2.18.6
pymysql==1.0.2
colorama==0.4.5
customtkinter==4.6.3
Pillow==9.2.0
cleo~=0.8.1
argon2-cffi~=21.3.0
password-validator==1.0